watchman: added script to install fsmonitor hook
Git status is slow in big repo. One of the best solution on the market
is Facebook's watchman which provide fsmonitor as a service. Microsoft
contributed the fsmonitor-watchman hook to upstream git that let user
leverage watchman for a faster git status.

git-care should detects whether user has watchman installed and
install/remove the hook properly.