commit-graph: Raised size-multiple from 2 to 4
For bigger repositories, its a good idea to have multiple commit_graph
chains file. But even at half the size, sometimes update and rewriting
the commit-graph could be too much.

Raising the size-multiple flag from default value of '2' to '4' should
make the top part of the graph to be a lot smaller, and cheaper to
refresh/rewrite.