Increased sleep interval for some jobs
Prefetch save downloaded objects as thin pack, thanks to `unpackLimit=1`
was set. This, however, runs faster than MIDX repack, so given the same
sleep interval for 2 jobs, Prefetch will be run more regularly.
Increasing the sleep interval for prefetch shall help balance packfiles' rate
of increase (due to prefetch) and rate of decrease (due to MIDX repack).

Pack Refs interval is subjectived to the rate of increase in branch/tag
in a remote repo. For our use case, the rate of branch/tag growth is not
too much of a concern for a 5mins interval (3-4 prefetch) hence increased
to 5mins.

Refresh Index job could have a race condition, so its best for us to
just increase it to reduce the chance of it being run when the user's
command were also trying to update the index.