Improved documentation
Added some documentation to reminds user about the watchman requirement
of the script.

Added a step counter to tests run and a '(slow)' tag to inform user
that the slowness might be expected, especially in commit-graph/Bloom
filter calculation and MIDX Repack.

Added doc to remind users that sleep interval(s) could be tuned from git
config of the repository.

Added doc to remind users that force kill is an option. This is relevant
when users might want to perform a quick reset.