Added expire_reflog job
Pre-fetch a lot could increase the size of your reflog.
This usually is cleaned by git-gc but because git-care turn off git-gc
to avoid git-repack call, reflog are left around

Implement expire_reflog job that cleans up reflog in the repository.
expire_reflog respects the expire duration so if you set git config in
the repo to gc.reflogExpire=never, logs are still not removed.